django/requirements.txt
django
django-redis
requests
gunicorn
psycopg2
psycopg2-binary
uvicorn[django]
python-dotenv